+++ request/data.go
@@ -1,0 +1,12 @@
+package request
+
+type Data struct {
+	Host     string
+	Path     string
+	Username string
+	Password string
+}
+
+func (data Data) Url() string {
+	return data.Host + data.Path
+}